% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{AIC.mle}
\alias{AIC.mle}
\title{Compute AIC for mle objects}
\usage{
\method{AIC}{mle}(object, finite = F)
}
\arguments{
\item{object}{a mle object.}

\item{finite}{logical; if true, applies correction for
finite sample size (e.g., Wagenmakers & Farrell, 2004).}
}
\description{
Compute AIC for mle objects
}
\details{
The correction for finite samples is \eqn{2K(K+1)/(n-K-1)}
  where K is the number of parameters and n is the number of
  observations.
}
\section{References}{

Wagenmakers, E. J., & Farrell, S. (2004). AIC model selection
  using Akaike weights. Psychonomic bulletin & review, 11,
  192-196.
}

