% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_functions.R
\name{akaike_weights}
\alias{akaike_weights}
\title{Compute Akaike/Schwarz weights}
\usage{
akaike_weights(ic_values, modelNames = NULL)
}
\arguments{
\item{ic_values}{a vector of AIC, BIC, or other information
criterion values.}

\item{modelNames}{an optional vector of labels for each model.}
}
\value{
A vector of the relative probabilities for each
  model.
}
\description{
Computes Akaike or Schwarz weights (e.g., Wagenmakers & Farrell,
2004) given a vector of information criterion values.
}
\section{References}{

Wagenmakers, E. J., & Farrell, S. (2004). AIC model selection
  using Akaike weights. Psychonomic bulletin & review, 11,
  192-196.
}
\examples{
data( ToothGrowth ) # Load in data
# Fit 2 regression models
m1 = lm( len ~ dose, data = ToothGrowth )
m2 = lm( len ~ dose + supp, data = ToothGrowth )
# Create vector of AIC values
ic = c( AIC(m1), AIC(m2) )
akaike_weights( ic )
# Include meaningful labels
modelNames = c( 'Dose', 'Dose+Supp' )
akaike_weights( ic, modelNames )

}

