% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{anova.mle}
\alias{anova.mle}
\title{Table of model comparisons using AIC and BIC}
\usage{
\method{anova}{mle}(..., finite = T, modelNames = NULL)
}
\arguments{
\item{...}{A set of mle objects.}

\item{finite}{logical; if true, applies the finite sample
correction to AIC calculations.}

\item{modelNames}{an optional character vector of labels
for the mle objects.}
}
\description{
Table of model comparisons using AIC and BIC
}

