% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{coef.mle}
\alias{coef.mle}
\title{Extract parameters from mle objects}
\usage{
\method{coef}{mle}(object)
}
\arguments{
\item{object}{a mle object.}
}
\description{
Extract parameters from mle objects
}

