\name{confint-methods}
\docType{methods}
\alias{confint-methods}
\alias{confint,ANY-method}
\alias{confint,profile.mle-method}
\alias{confint,mle-method}
\title{Methods for Function confint in Package `mle'}
\description{Generate confidence intervals}
\section{Methods}{\describe{

\item{object = "ANY"}{Generic function}

\item{object = "profile.mle"}{Generate confidence intervals based on
  likelihood profile.}

\item{object = "mle"}{First generate profile and then confidence
  intervals as above.}
}}
\keyword{methods}
