% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{confint.mle}
\alias{confint.mle}
\title{Extract confidence intervals from mle objects}
\usage{
\method{confint}{mle}(object)
}
\arguments{
\item{object}{a mle object.}
}
\description{
Extract confidence intervals from mle objects
}

