% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_functions.R
\name{create_mle_fn}
\alias{create_mle_fn}
\title{Create a MLE function}
\usage{
create_mle_fn(f)
}
\arguments{
\item{f}{a function of the form \code{f(dat,prm)}, where
\code{dat} is the data to be fitted and \code{prm} is
a vector of parameters.}
}
\value{
A new function that takes 4 variables:
  \describe{
    \item{\code{dat}}{The data to be fitted}
    \item{\code{prm}}{A vector of parameters}
    \item{\code{sum}}{A logical value for summing the
      log-likeihoods}
    \item{\code{priors}}{A nuisance variable included for
      compatibility}
  }
}
\description{
Given a simple log-likelhood function, produces a new
function that can be passed into the \code{MLE} function.
}
\examples{
f = function(dat,prm) dnorm( dat, prm[1], exp( prm[2] ), log = T )
mle_fn = create_mle_fn(f)
MLE( rnorm( 100 ), mle_fn, c(0,1) )

}

