% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{deviance.mle}
\alias{deviance.mle}
\title{Extract model deviance from mle objects}
\usage{
\method{deviance}{mle}(object)
}
\arguments{
\item{object}{a mle object.}
}
\description{
Extract model deviance from mle objects
}

