% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{logLik.mle}
\alias{logLik.mle}
\title{Extract log-likelihood from mle objects}
\usage{
\method{logLik}{mle}(object)
}
\arguments{
\item{object}{a mle object.}
}
\description{
Extract log-likelihood from mle objects
}

