% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_exp.R
\name{mle_exp}
\alias{mle_exp}
\title{Estimasi Maximum Likelihood untuk Distribusi Eksponensial}
\usage{
mle_exp(x, na.rm = TRUE)
}
\arguments{
\item{x}{Sebuah vektor numerik berisi data observasi. Nilai harus non-negatif.}

\item{na.rm}{Sebuah nilai logis yang menunjukkan apakah nilai `NA` harus
dihilangkan sebelum perhitungan. Standarnya adalah `TRUE`.}
}
\value{
Sebuah vektor numerik dengan nama yang berisi satu elemen:
\itemize{
  \item `lambda_hat`: Estimasi parameter *rate* (\eqn{\hat{\lambda}}).
}
}
\description{
Fungsi ini menghitung *Maximum Likelihood Estimator* (MLE) untuk parameter
*rate* (\eqn{\lambda}) dari distribusi Eksponensial berdasarkan sebuah vektor numerik.
}
\details{
Formula yang digunakan adalah estimator MLE untuk parameter rate:
\eqn{\hat{\lambda} = \frac{1}{\bar{x}} = \frac{n}{\sum_{i=1}^{n}x_i}}
}
\examples{
# Bangkitkan data sampel dari distribusi Eksponensial
set.seed(456)
data_exp <- rexp(100, rate = 2)

# Hitung MLE
mle_exp(data_exp)

# Contoh dengan data yang mengandung nilai NA dan nol
data_with_na_and_zero <- c(data_exp, NA, 0)
mle_exp(data_with_na_and_zero, na.rm = TRUE)
}
