% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_functions.R
\name{mle_fn_template}
\alias{mle_fn_template}
\title{Template a log-likelihood function}
\usage{
mle_fn_template(version = 0)
}
\arguments{
\item{version}{Indicates the type of template to return.}
}
\value{
Prints a template for a function to the console that
  may then be copied and modified for the user's needs.
}
\description{
Generates a template for a function to compute the log-likelihood
or sum of log-likelihoods for data given a vector of parameters.
This function can then be passed into the \code{MLE} function.
}
\examples{
mle_fn_template()

}

