\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,ANY,ANY-method}
\alias{plot,profile.mle,missing-method}
\title{Methods for Function plot in Package `mle'}
\description{Plot objects}
\section{Methods}{\describe{

\item{x = "ANY", y = "ANY"}{Generic function}

\item{x = "profile.mle", y = "missing"}{Plot likelihood profiles for
  \code{x}}
}}
\keyword{methods}
