% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.anova.mle}
\alias{print.anova.mle}
\title{Print model comparisons for mle objects}
\usage{
\method{print}{anova.mle}(out)
}
\arguments{
\item{out}{output from \code{anova} call on a set of mle
objects.}
}
\description{
Print model comparisons for mle objects
}

