% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.summary.mle}
\alias{print.summary.mle}
\title{Print summary measures for mle objects}
\usage{
\method{print}{summary.mle}(res)
}
\arguments{
\item{res}{output from summary of a mle object.}
}
\description{
Print summary measures for mle objects
}

