\name{MLE of distributions defined in the (0, 1) interval}
\alias{prop.mle}
\title{
MLE of distributions defined in the (0, 1) interval
}

\description{
MLE of distributions defined in the (0, 1) interval.
}

\usage{
prop.mle(x, distr = "beta", tol = 1e-07, maxiters = 50)
}

\arguments{
\item{x}{
A numerical vector with proportions, i.e. numbers in (0, 1) (zeros and ones are not allowed).
}
\item{distr}{
The distribution to fit. "beta" stands for the beta distribution, "logitnorm" is the logistic normal,
"unitweibull" is the unit-Weibull and the "sp" is the standard power distribution, "ibeta" is the
inflated beta, (0-inflated or 1-inflated, depending on the data), "hsecant01" stands for the hyper-secant,
"kumar" is the Kumaraswamy, "simplex" is the simplex distribution, "zil" is the zero inflated logistic normal,
"cbern" is the continuous Bernoulli distribution and "logisbeta" is the logistic-beta distribution, and
"garcsine" is the generalised arcsine distribution.
}
\item{tol}{
The tolerance level up to which the maximisation stops.
}
\item{maxiters}{
The maximum number of iterations to implement.
}
}

\details{
Maximum likelihood estimation of the parameters of the beta distribution is performed via Newton-Raphson. The distributions and hence the functions does not accept zeros. "logitnorm" fits the logistic normal, hence no nwewton-Raphson is required and the "hypersecant01" uses the golden ratio search as is it faster than the Newton-Raphson (less calculations). The distributions included are the Kumaraswamy, zero inflated logistic normal, simplex, unit Weibull and continuous Bernoulli and standard power. Instead of maximising the log-likelihood via a numerical optimiser we have used a Newton-Raphson algorithm which is faster. See wikipedia for the equations to be solved.
}

\value{
A list including:
\item{iters}{
The number of iterations required by the Newton-Raphson.
}
\item{loglik}{
The value of the log-likelihood.
}
\item{param}{
The estimated parameters. In the case of "hypersecant01.mle" this is called
"theta" as there is only one parameter.
}
}

\references{
Kumaraswamy P. (1980). A generalized probability density function for double-bounded
random processes. Journal of Hydrology 46(1-2): 79--88.

Jones M.C. (2009). Kumaraswamy's distribution: A beta-type distribution with some
tractability advantages. Statistical Methodology, 6(1): 70--81.

J. Mazucheli, A. F. B. Menezes, L. B. Fernandes, R. P. de Oliveira and M. E. Ghitany (2020).
The unit-Weibull distribution as an alternative to the Kumaraswamy distribution for the modeling of quantiles conditional on covariates. Journal of Applied Statistics, 47(6): 954--974.

Leemis L.M. and McQueston J.T. (2008). Univariate Distribution Relationships.
The American Statistician, 62(1): 45--53.

You can also check the relevant wikipedia pages.
}

\author{
Michail Tsagris and Sofia Piperaki.

R implementation and documentation: Michail Tsagris \email{mtsagris@uoc.gr} and Sofia Piperaki  \email{sofiapip23@gmail.com}.
}

%\note{
%%  ~~further notes~~
%}

\seealso{
\code{ \link{colprop.mle}, \link{comp.mle}
}
}

\examples{
x <- rbeta(1000, 1, 4)
prop.mle(x, distr = "beta")
}
