% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.mle}
\alias{summary.mle}
\title{Computes summary measures for mle objects}
\usage{
\method{summary}{mle}(object, crit = 0)
}
\arguments{
\item{object}{a mle object..}

\item{crit}{comparison value(s) used in the Wald test.}
}
\description{
Computes summary measures for mle objects
}

