% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_functions.R
\name{tran_par}
\alias{tran_par}
\title{Transform a vector of parameters}
\usage{
tran_par(prm, type, reverse = F)
}
\arguments{
\item{prm}{a vector of parameters.}

\item{type}{a vector of the type of transformation to apply
to each parameter, where...
\itemize{
  \item{0 = no transformation;}
  \item{1 = exponentiation;}
  \item{2 = the logistic transform;}
  \item{3 = the inverse.}
}}

\item{reverse}{a logical value; if true, reverses the transformation
specified by \code{type}.}
}
\value{
A vector of transformed or reversed-transformed parameters.
}
\description{
A function to transforms or reverse transformations of
parameter estimates, which is particularly useful to
bound estimates to be positive or restricted to a set
range.
}
\examples{
prm = c( 1, -1, 0, .5 )
prm_star = tran_par( prm, 0:3 )
prm_star
prm == tran_par( prm_star, 0:3, reverse = T )

}

