% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{vcov.mle}
\alias{vcov.mle}
\title{Extract the variance-covariance matrix from mle objects}
\usage{
\method{vcov}{mle}(object)
}
\arguments{
\item{object}{a mle object.}
}
\description{
Extract the variance-covariance matrix from mle objects
}

