\name{MLE-package}
\alias{MLE-package}
\alias{MLE}
\docType{package}
\title{
\packageTitle{MLE}
}
\description{
\packageDescription{MLE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MLE}
\packageIndices{MLE}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{MLE}

Maintainer: \packageMaintainer{MLE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
