% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_functions.R
\name{MLE}
\alias{MLE}
\title{A convenience function to carry out maximum likelihood estimation
using either \code{optim} or \code{nlm}.}
\usage{
MLE(dat, mle_fn, start, priors = NULL, gr_fn = NULL, method = NULL,
  control = list(), hessian = T, alpha = 0.9544997, emStop = 20,
  nRep = 1, parNames = NULL)
}
\arguments{
\item{dat}{the data to be fitted; can be in a variety of forms
as long as it is compatible with the \code{mle_fn} function
that is passed in.}

\item{mle_fn}{a function to compute the sum of the log-likelihoods
for a vector of parameters \code{prm}, the data \code{dat}, and
control parameters \code{sum} and \code{priors}.}

\item{start}{either a vector of initial starting values for the
parameters, or a function (which takes no parameters) used to
generate a dispersed set of starting values.}

\item{gr_fn}{an optional function to compute the gradients for the
parameters being estimated. See \code{\link{optim}}.}

\item{method}{the optimization algorithm to use. See
\code{\link{optim}}. If \code{NULL}, a step-by-step approach
is used, where viable estimates and standard errors are sought
by first trying the 'BFGS' method, then the 'BFGS' method using
initial estimates from a 'Nelder-Mead' approach, then the 'nlm'
approach, and finally the 'Nelder-Mead' algorithm.}

\item{control}{a named list where options controlling the
optimization can be specified. See \code{\link{optim}} and
\code{\link{nlm}}. The lower and upper limits for the 'Brent'
method can also be specified in this list.}

\item{hessian}{a logical value; if true, the hessian matrix is
estimated, allowing the derivation of standard errors and
confidence intervals.}

\item{alpha}{the width of the interval for the confidence intervals
around parameters. Defaults to +/- two standard deviations.}

\item{emStop}{the number of iterations used when attempting to
generate viable starting values.}

\item{nRep}{the number of times to repeat estimation using
dispersed starting values. Only applicable when a function to
generate starting values is provided.}

\item{parNames}{an optional character vector giving a set of labels
for the parameters being estimated.}
}
\value{
Forthcoming.
}
\description{
A convenience function to carry out maximum likelihood estimation
using either \code{optim} or \code{nlm}.
}

